test_project/requirements.txt
amqp==5.1.1
ansicon==1.89.0
arrow==1.2.3
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
billiard==4.1.0
blessed==1.20.0
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.4
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==4.2.3
django-cors-headers==4.1.0
django-picklefield==3.1
django-q==1.3.9
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
frozenlist==1.3.3
idna==3.4
jinxed==1.2.0
kombu==5.3.1
multidict==6.0.4
prompt-toolkit==3.0.39
PyJWT==2.7.0
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
rest-framework-simplejwt==0.0.2
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.3
vine==5.0.0
wcwidth==0.2.6
yarl==1.9.2
